+++ stories/form-control/FormControl.component.ts
@@ -1,15 +1,19 @@
 /* eslint-disable no-console */
 import { Component } from '@angular/core';
-import { FormBuilder } from '@angular/forms';
+import { FormBuilder, FormControl } from '@angular/forms';
+import { apiKey } from 'stories/Settings';
 
 @Component({
-  templateUrl: './FormControl.component.html'
+  selector: 'form-control',
+  templateUrl: './FormControl.component.html',
 })
 export class FormControlComponent {
-  public formControl = this.formBuilder.control<string | null>(null);
+  public apiKey = apiKey;
+  public formControl: FormControl<string | null>;
 
   // eslint-disable-next-line @typescript-eslint/no-parameter-properties
   public constructor(private readonly formBuilder: FormBuilder) {
+    this.formControl = this.formBuilder.control<string | null>(null);
     this.formControl.valueChanges.subscribe(console.log);
     this.formControl.setValue('<p>Initial value</p>');
     // Console log should be triggered just once