+++ tinymce-angular-component/src/test/ts/alien/InitTestEnvironment.ts
@@ -1,7 +1,6 @@
 import 'core-js/features/reflect';
 import 'zone.js';
-import 'zone.js/dist/fake-async-test.js';
-import 'zone.js/dist/zone';
+import 'zone.js/plugins/fake-async-test';
 
 import { TestBed } from '@angular/core/testing';
 import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';