+++ tinymce-angular-component/src/test/ts/browser/NgModelTest.ts
@@ -11,10 +11,9 @@ import { VersionLoader } from '@tinymce/miniature';
 import { SugarElement } from '@ephox/sugar';
 
 import { EditorModule, EditorComponent } from '../../../main/ts/public_api';
+import { Version } from '../../../main/ts/editor/editor.component';
 
 UnitTest.asynctest('NgModelTest', (success, failure) => {
-  // tslint:disable:max-classes-per-file
-
   class Base {
     public content: string | undefined;
   }
@@ -72,7 +71,7 @@ UnitTest.asynctest('NgModelTest', (success, failure) => {
     context.fixture.detectChanges();
   });
 
-  const sTestVersion = (version: '4' | '5' | '6') => VersionLoader.sWithVersion(version, GeneralSteps.sequence([
+  const sTestVersion = (version: Version) => VersionLoader.sWithVersion(version, GeneralSteps.sequence([
     Log.chainsAsStep('', 'should be pristine, untouched, and valid initially', [
       cSetupEditorWithNgModel(),
       cAssertNgModelState('valid', true),
@@ -146,6 +145,7 @@ UnitTest.asynctest('NgModelTest', (success, failure) => {
   Pipeline.async({}, [
     sTestVersion('4'),
     sTestVersion('5'),
-    sTestVersion('6')
+    sTestVersion('6'),
+    sTestVersion('7')
   ], success, failure);
 });